+++ shinywidgets/__init__.py
@@ -10,7 +10,6 @@ from ._output_widget import output_widget
 from ._render_widget import (
     render_altair,
     render_bokeh,
-    render_leaflet,
     render_plotly,
     render_pydeck,
     render_widget,
@@ -22,7 +21,6 @@ __all__ = (
     "render_widget",
     "render_altair",
     "render_bokeh",
-    "render_leaflet",
     "render_plotly",
     "render_pydeck",
     # Reactive read second