+++ shinywidgets/_shinywidgets.py
@@ -1,12 +1,9 @@
 from __future__ import annotations
 
 import copy
-import importlib
 import json
 import os
-import tempfile
-import warnings
-from typing import Any, Optional, Sequence, Tuple, Union, cast
+from typing import Any, Optional, Sequence, Union, cast
 from uuid import uuid4
 from weakref import WeakSet
 
@@ -28,7 +25,8 @@ from shiny.session import get_current_session, require_active_session
 from ._as_widget import as_widget
 from ._cdn import SHINYWIDGETS_CDN_ONLY, SHINYWIDGETS_EXTENSION_WARNING
 from ._comm import BufferType, ShinyComm, ShinyCommManager
-from ._dependencies import require_dependency, widget_pkg
+from ._dependencies import require_dependency
+from ._utils import is_instance_of_class, package_dir
 
 __all__ = (
     "register_widget",
@@ -149,9 +147,10 @@ COMM_MANAGER = ShinyCommManager()
 # Reactivity
 # --------------------------------------
 
-
-# TODO-barret; Make method on RenderWidget base
 def reactive_read(widget: Widget, names: Union[str, Sequence[str]]) -> Any:
+    """
+    Reactively read a widget trait
+    """
     reactive_depend(widget, names)
     if isinstance(names, str):
         return getattr(widget, names)
@@ -165,7 +164,7 @@ def reactive_depend(
     type: str = "change",
 ) -> None:
     """
-    Reactively read a Widget's trait(s)
+    Take a reactive dependency on a widget trait
     """
 
     try:
@@ -195,10 +194,12 @@ def reactive_depend(
     ctx.on_invalidate(_)
 
 
-# TODO-barret; Proposal: Remove `register_widget()` in favor of `@render_widget`
 def register_widget(
     id: str, widget: Widget, session: Optional[Session] = None
 ) -> Widget:
+    """
+    Deprecated. Use @render_widget instead.
+    """
     if session is None:
         session = require_active_session(session)
 
@@ -210,90 +211,6 @@ def register_widget(
         return w
 
     return w
-
-
-def set_layout_defaults(widget: Widget) -> Tuple[Widget, bool]:
-    # If we detect a user specified height on the widget, then don't
-    # do filling layout (akin to the behavior of output_widget(height=...))
-    fill = True
-
-    if not isinstance(widget, DOMWidget):
-        return (widget, fill)
-
-    # Do nothing for "input-like" widgets (e.g., ipywidgets.IntSlider())
-    if getattr(widget, "_model_module", None) == "@jupyter-widgets/controls":
-        return (widget, False)
-
-    layout = widget.layout  # type: ignore
-
-    # If the ipywidget Layout() height is set to something other than "auto", then
-    # don't do filling layout https://ipywidgets.readthedocs.io/en/stable/examples/Widget%20Layout.html
-    if isinstance(layout, Layout):
-        if layout.height is not None and layout.height != "auto":  # type: ignore
-            fill = False
-
-    pkg = widget_pkg(widget)
-
-    # Plotly provides it's own layout API (which isn't a subclass of ipywidgets.Layout)
-    if pkg == "plotly":
-        from plotly.graph_objs import Layout as PlotlyLayout  # pyright: ignore
-
-        if isinstance(layout, PlotlyLayout):
-            if layout.height is not None:  # pyright: ignore[reportUnknownMemberType]
-                fill = False
-            # Default margins are also way too big
-            layout.template.layout.margin = dict(  # pyright: ignore
-                l=16, t=32, r=16, b=16
-            )
-            # Unfortunately, plotly doesn't want to respect the top margin template,
-            # so change that 60px default to 32px
-            if layout.margin["t"] == 60:  # pyright: ignore
-                layout.margin["t"] = 32  # pyright: ignore
-
-    widget.layout = layout
-
-    # altair, confusingly, isn't setup to fill it's Layout() container by default. I
-    # can't imagine a situation where you'd actually want it to _not_ fill the parent
-    # container since it'll be contained within the Layout() container, which has a
-    # full-fledged sizing API.
-    if pkg == "altair":
-        import altair as alt
-
-        # Since as_widget() has already happened, we only need to handle JupyterChart
-        if isinstance(widget, alt.JupyterChart):
-            if isinstance(
-                widget.chart,  # pyright: ignore[reportUnknownMemberType]
-                alt.ConcatChart,
-            ):
-                # Throw warning to use ui.layout_column_wrap() instead
-                warnings.warn(
-                    "Consider using shiny.ui.layout_column_wrap() instead of alt.concat() "
-                    "for multi-column layout (the latter doesn't support filling layout)."
-                )
-            else:
-                widget.chart = widget.chart.properties(width="container", height="container")  # type: ignore
-
-    return (widget, fill)
-
-
-# similar to base::system.file()
-def package_dir(package: str) -> str:
-    with tempfile.TemporaryDirectory():
-        pkg_file = importlib.import_module(".", package=package).__file__
-        if pkg_file is None:
-            raise ImportError(f"Couldn't load package {package}")
-        return os.path.dirname(pkg_file)
-
-
-def is_instance_of_class(
-    x: object, class_name: str, module_name: Optional[str] = None
-) -> bool:
-    typ = type(x)
-    res = typ.__name__ == class_name
-    if module_name is None:
-        return res
-    else:
-        return res and typ.__module__ == module_name
 
 
 # It doesn't, at the moment, seem feasible to establish a comm with statically rendered widgets,